chore: mv deep equal into shared
