+++ apps/emdash-desktop/src/main/core/settings/utils.ts
@@ -1,4 +1,4 @@
-import { isDeepEqual } from '@emdash/core/lib';
+import { isDeepEqual } from '@emdash/shared';
 
 export function isPlainObject(v: unknown): v is Record<string, unknown> {
   return typeof v === 'object' && v !== null && !Array.isArray(v);
