+++ packages/core/src/lib/deep-equal.ts
@@ -1,32 +1,1 @@
-/**
- * Structural equality for plain JSON-style data: primitives, arrays, and plain objects.
- */
-export function isDeepEqual(a: unknown, b: unknown): boolean {
-  if (Object.is(a, b)) return true;
-
-  if (Array.isArray(a) || Array.isArray(b)) {
-    if (!Array.isArray(a) || !Array.isArray(b)) return false;
-    if (a.length !== b.length) return false;
-    for (let i = 0; i < a.length; i++) {
-      if (!isDeepEqual(a[i], b[i])) return false;
-    }
-    return true;
-  }
-
-  if (!isPlainRecord(a) || !isPlainRecord(b)) return false;
-
-  const keysA = Object.keys(a);
-  const keysB = Object.keys(b);
-  if (keysA.length !== keysB.length) return false;
-  for (const key of keysA) {
-    if (!Object.hasOwn(b, key)) return false;
-    if (!isDeepEqual(a[key], b[key])) return false;
-  }
-  return true;
-}
-
-function isPlainRecord(value: unknown): value is Record<string, unknown> {
-  if (typeof value !== 'object' || value === null) return false;
-  const prototype = Object.getPrototypeOf(value);
-  return prototype === Object.prototype || prototype === null;
-}
+export { isDeepEqual } from '@emdash/shared';
