+++ packages/core/src/lib/index.ts
@@ -1,4 +1,4 @@
-export { isDeepEqual } from './deep-equal';
+export { isDeepEqual } from '@emdash/shared';
 export { Emitter } from './emitter';
 export { KeyedMutex } from './keyed-mutex';
 export { consoleLogger, noopLogger, type Logger } from './logger';
