+++ packages/shared/src/index.ts
@@ -8,4 +8,5 @@ export {
   type Ok,
   type Result,
 } from './result';
+export { isDeepEqual } from './deep-equal';
 export type { IDisposable, IInitializable, ILifecycle, Lease, Unsubscribe } from './lifecycle';
